+++ overlookcam-relay/run.sh
@@ -3,7 +3,7 @@ set -euo pipefail
 
 export OVERLOOKCAM_INSTALLATION_STATE="/data/installation.json"
 export OVERLOOKCAM_RELAY_PLATFORM="home-assistant"
-export OVERLOOKCAM_RELAY_VERSION="0.2.6"
+export OVERLOOKCAM_RELAY_VERSION="0.2.7"
 export PUSH_LABELS="person,dog,cat"
 
 bashio::log.info "Starting OverlookCam Relay manager"